refactor(test-stubs): extract common values from NormalQueryComponent stub
These values are going to be used by other stubs as well.
