+++ __tests__/stubs/NormalQueryComponent.tsx
@@ -1,23 +1,8 @@
 import React from 'react';
-import { gql } from '@apollo/client';
 
 import useQueryWithTTL from '../../src/use-query-with-ttl';
 
-export const Query = gql`
-  query Todos {
-    todoList {
-      id
-      description
-    }
-  }
-`;
-
-const Constants = {
-  hookParams: {
-    query: Query,
-    ttl: 100,
-  },
-};
+import { Constants } from './shared';
 
 interface NormalQueryComponentProps {
   onCompleted?: () => void;
